+++ src/components/Colors/styles.stories.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Container = styled.div<any>`
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  gap: 1rem;
+`;